hotfix: useSWR revalidations of the page using broken compare (#198)
When adding metronome we also updated [Song] to have three new functions
attached to it. The builtin compare for useSWR could no longer determine that
the song hadn't changed, because the object identity of the new functions were
different.

While we should think about fixing that, a better fix for this instance would
be to use useSWRImmutable instead, since we can assume songdata will never
change during the lifetome of a session.